修复一个奇怪的问题

+++ api-admin/src/main/java/com/lmxdawn/api/admin/service/carouel/CarouselService.java
@@ -8,6 +8,6 @@ public interface CarouselService {
      boolean Insert(Carousel entity);
     boolean Update(Carousel entity);
     List<Carousel> GetAll();
-    boolean Delete(long id);
+    boolean Delete(int id);
 
 }
